Added more file time debugging

+++ Libraries/Common/FileSystem.cpp
@@ -147,14 +147,20 @@ int CheckFileTime(StringParam dest, StringParam source)
 
   if (destTime < sourceTime)
   {
-    DebugFileTime(
-        "Source '%s' (%d) is newer than destination '%s' (%d)\n", source.c_str(), sourceTime, dest.c_str(), destTime);
+    DebugFileTime("Source '%s' (%d) is newer than destination '%s' (%lld)\n",
+                  source.c_str(),
+                  sourceTime,
+                  dest.c_str(),
+                  (long long)destTime);
     return -1;
   }
   if (destTime > sourceTime)
   {
-    DebugFileTime(
-        "Source '%s' (%d) is older than destination '%s' (%d)\n", source.c_str(), sourceTime, dest.c_str(), destTime);
+    DebugFileTime("Source '%s' (%d) is older than destination '%s' (%lld)\n",
+                  source.c_str(),
+                  sourceTime,
+                  dest.c_str(),
+                  (long long)destTime);
     return +1;
   }
